Move to conda environment. (#36)
This PR synchronizes the image with the aiida-core where conda is 
the default python environment. It merges all pip installations into
one cleans up the code. Also, it removes installation of
fastentrypoints package that is not needed anymore.

The `.travis` file has been updated to show the log of container launch
and the status of aiida daemon.

In addition to that 3 new packages are installed through `conda install`:
rdkit, openbabel, dscribe.

Conda environment activation has been added to prepare-aiidalab.sh 
script.