Install jupyter in 'root' (not conda) environment. (#64)
Previously, everything was installed in the base conda environment,
however, this created some problems in coupling jupyter with AiiDA
which requires tornado to be lower than 5.0.0. To fix this problem
this PR installs jupyter in a 'root' environment, but configures it
to use the ipykernel from the base conda environment.