Fix Python version in build.json (#326)
This was an omission in #318. The Python version specified in build.json is used to automatically generate the image tags on Dockerhub. So currently, for example, the image aiidalab/full-stack:python-3.9.4 actually corresponds to python-3.9.13.
The default BASE in `stack/base/Dockerfile` is set in docker-bake.hcl for building the image. 
Replace the `ARG BASE` with using the arg from the environment should be fine and good to have only one place to set the version.