Using `notebook` CMD to run the jupyter stack (#331)
The new jupyter/minimal-notebook image uses lab CMD to run the jupyter server, which was introduced by jupyter/docker-stacks#1575. It provides DOCKER_STACKS_JUPYTER_CMD to set how to run it.
The lab (full-stack) is for AiiDAlab only and should start with notebook mode. I pin it as an environment variable to start the jupyter backend.
Since we have the new jupyterhub, the `NOTEBOOK_ARGS` can be used to set the notebook arguments.