Fix: Add missing path for $HOME/.local/bin (#325)
Add `~/.local/bin` to PATH where the dependencies get installed via pip

In the old Docker stack image, the conda environment in `/opt/conda` was not writeable by the aiida user. As a side-effect of that, when the AiiDAlab app was installed via pip, its declared dependencies (and the app itself) was installed in the `/home/aiida/.local` directory. These dependencies then persisted between different container invocation since they were stored in the home volume.

In the new full-stack image, the conda environment is writeable by the jovyan user, hence pip defaults to the system-wide installation, in this case /opt/conda. Therefore, the installed dependencies disappear when the container exits.
In the following invocation, the Apps won't load due to missing dependencies.

The fix is to explicitly create a folder in the container for python packages in the `~/.local/bin`.

Co-authored-by: Jusong Yu <jusong.yeu@gmail.com>