Add .vimrc to home folder for nocompatible vim (#339)
once a personal initialization file ~/.vimrc exists, Vim automatically turns on 'nocompatible' mode.

Compatible mode means compatibility to venerable old vi. When you :set compatible, all the enhancements and improvements of Vi Improved are turned off. It's not recommended to do this, but some systems provide (mostly for backward compatibility with old Unix systems) a vi command that is implemented with Vim in compatible mode.