Configure custom pytest marker for integration tests
We're using a custom 'integration' pytest marker
for testing that we can successfully install AWB
and QeApp into the image.

Pytest currently emits a warning
because it doesn't recognize this custom marker.

"""
PytestUnknownMarkWarning: Unknown pytest.mark.integration - is this a typo?
You can register custom marks to avoid this warning - for details, see
https://docs.pytest.org/en/stable/how-to/mark.html
"""

Here we silence the warning by explicitly defining
the markers in pytest.ini file. We also configure pytest
to error out if an unknown marker is found in the future,
as opposed to only emitting a warning.