Configure custom pytest marker for integration tests (#349)
We're using a custom `integration` pytest marker for testing that we can successfully install AWB
and QeApp into the image. Pytest currently emits a warning because it doesn't recognize this custom
marker.

Here we silence the warning by explicitly defining the markers in the `pytest.ini` file. We also configure pytest
to error out if an unknown marker is found in the future (instead of only emitting a warning).