Move rc configures in home folder to before-notebook (#353)
If the container is started by the spawner and the home is bind mounted the content of the container will be hidden away. 
It leads to the `.bashrc` in HOME won't be set properly. I generate the .bashrc, .profile, and .bash_logout from `/etc/skel/` folder and move the `.bashrc` modification for singlesshagent into the before-notebook script.

For more details about how docker and k8s manage the bind mount:
- https://docs.docker.com/storage/bind-mounts/#mount-into-a-non-empty-directory-on-the-container
- https://stackoverflow.com/questions/59979121/kubernetes-persistent-volumes-deleting-initial-files