Pin major PostgreSQL version in the image
We're only pinning the major version since
minor versions shouldn't require DB migrations.

I am also adding the postgresql-{version} docker tag to stable releases,
in the same way we tag the images with Python and AiiDA versions.
That should make it easy for users to avoid migrations if they need to.