Test full-stack and lab (#385)
The CI tests were run on `aiida-with-services` and lab images by using `pytest` params to control. Based on which image is tested, different tests are running. The AiiDAlab related tests are skipped since the image `aiida-with-services` does not include aiidalab. PostgreSQL installation tests are skipped for lab image test because the services are launched as a separate container in docker-compose.

It simply increases the complicity of testing, so in this PR I test for full-stack and lab which all have aiidalab installed therefore no corresponding tests need to be skipped.