Using reusable CI workflows for docker stack build (#390)
This is a rearrangement of the docker build CI. The CI was run to build everything (every target with two architecture, that is 8 things to build) in one action which makes the CI takes very long and hard to debug. The tests also happened within one action for all the docker stacks.

I adopt the way how Jupyter community manages their CI test, https://github.com/jupyter/docker-stacks. The idea is to use GitHub reusable workflow without duplicating the CI workflows.

The final outcome of the build didn't change too much, for every new PR, once it successfully build and tests passed, the docker image with tag pr-xxx is uploaded to GitHub container registry https://github.com/orgs/aiidalab/packages. https://github.com/aiidalab/aiidalab-docker-stack/pkgs/container/full-stack is the full-stack image of this PR. For different commits of the PR, it is distinguished by a new sha-xxx tag. The pr-xx is pointed to the latest built images(and tests passed, this is different from the previous implementation where the unit tests are run after uploading to ghcr.

A new thing is for every image build, the image is exported to the workflow artifacts and kept there for 3 days, this can be downloaded and tested locally. The artifacts will be uploaded even if the unit tests are failed.