Implement pre-installed QE components for demo server (#402)
A new image is brought into the stack which includes the pre-installed Quantum ESPRESSO from conda forge to be run from localhost. 
The dependencies are installed from Dockerfile while the `aiidalab-qe` need to be installed the first time container started since the app should be installed in the user home which only created after the container started. Similar for the QE from conda forge, the Dockerfile creates the virtual env and installs the QE to the system folder and when the container started the first time it is copied to the user's home if the folder not existed. 

This image is designed as the image for the AiiDAlab demo server in the materials cloud so that new users can have a quick taste of AiiDAlab and run simulations with QeApp.