Pin jupyter-notebook version (#469)
In aiidalab/aiidalab-qe#640, we were debugging an issue with a stalled pip installation, 
during which pip started backtracking and looking at old versions of
some of the core Jupyter libraries, most notably Jupyter notebook (notebook package). 
This is definitely not desirable, and if pip was able to finish, it would likely break the container.

We've been having discussions in the past about pinning some jupyter-related packages to prevent this scenario from happening. Here we pin at least the notebook package. 
We definitely don't want users to be able to downgrade the version, 
and we also don't want them to upgrade it, since we know that v7 will
not work as it is a completely different architecture.