Fix build (#538)
The set-output step started failing again due to #493,
because the bake-action metadata output is too big to parse.

This commit hopefully fixes this issue for good:
We first write the metadata output to a file, 
which is then read in the JS script in the set-output step. 
Perhaps in the future there will be event a better solution,
see this request: docker/bake-action#178 (comment)